refactor: Names 생성을 NameCreator로 분리
